+++ src/typeguard/_checkers.py
@@ -32,7 +32,7 @@ from typing import (
 )
 from unittest.mock import Mock
 
-from ._config import ForwardRefPolicy, global_config
+from ._config import ForwardRefPolicy, TypeCheckConfiguration
 from ._exceptions import TypeCheckError, TypeHintWarning
 from ._memo import TypeCheckMemo
 from ._utils import evaluate_forwardref, get_type_name, qualified_name
@@ -75,7 +75,7 @@ else:
     from typing_extensions import ParamSpec
 
 TypeCheckerCallable: TypeAlias = Callable[
-    [Any, Any, Tuple[Any, ...], TypeCheckMemo], Any
+    [Any, Any, Tuple[Any, ...], TypeCheckMemo, TypeCheckConfiguration], Any
 ]
 TypeCheckLookupCallback: TypeAlias = Callable[
     [Any, Tuple[Any, ...], Tuple[Any, ...]], Optional[TypeCheckerCallable]
@@ -143,7 +143,11 @@ BINARY_MAGIC_METHODS = {
 
 
 def check_callable(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not callable(value):
         raise TypeCheckError("is not callable")
@@ -199,7 +203,11 @@ def check_callable(
 
 
 def check_mapping(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if origin_type is Dict or origin_type is dict:
         if not isinstance(value, dict):
@@ -213,25 +221,27 @@ def check_mapping(
     if args:
         key_type, value_type = args
         if key_type is not Any or value_type is not Any:
-            samples = global_config.collection_check_strategy.iterate_samples(
-                value.items()
-            )
+            samples = config.collection_check_strategy.iterate_samples(value.items())
             for k, v in samples:
                 try:
-                    check_type_internal(k, key_type, memo)
+                    check_type_internal(k, key_type, memo, config)
                 except TypeCheckError as exc:
                     exc.append_path_element(f"key {k!r}")
                     raise
 
                 try:
-                    check_type_internal(v, value_type, memo)
+                    check_type_internal(v, value_type, memo, config)
                 except TypeCheckError as exc:
                     exc.append_path_element(f"value of key {k!r}")
                     raise
 
 
 def check_typed_dict(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     declared_keys = frozenset(origin_type.__annotations__)
     if hasattr(origin_type, "__required_keys__"):
@@ -254,62 +264,78 @@ def check_typed_dict(
         argvalue = value.get(key, _missing)
         if argvalue is not _missing:
             try:
-                check_type_internal(argvalue, argtype, memo)
+                check_type_internal(argvalue, argtype, memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"value of key {key!r}")
                 raise
 
 
 def check_list(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isinstance(value, list):
         raise TypeCheckError("is not a list")
 
     if args and args != (Any,):
-        samples = global_config.collection_check_strategy.iterate_samples(value)
+        samples = config.collection_check_strategy.iterate_samples(value)
         for i, v in enumerate(samples):
             try:
-                check_type_internal(v, args[0], memo)
+                check_type_internal(v, args[0], memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"item {i}")
                 raise
 
 
 def check_sequence(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isinstance(value, collections.abc.Sequence):
         raise TypeCheckError("is not a sequence")
 
     if args and args != (Any,):
-        samples = global_config.collection_check_strategy.iterate_samples(value)
+        samples = config.collection_check_strategy.iterate_samples(value)
         for i, v in enumerate(samples):
             try:
-                check_type_internal(v, args[0], memo)
+                check_type_internal(v, args[0], memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"item {i}")
                 raise
 
 
 def check_set(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isinstance(value, AbstractSet):
         raise TypeCheckError("is not a set")
 
     if args and args != (Any,):
-        samples = global_config.collection_check_strategy.iterate_samples(value)
+        samples = config.collection_check_strategy.iterate_samples(value)
         for v in samples:
             try:
-                check_type_internal(v, args[0], memo)
+                check_type_internal(v, args[0], memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"[{v}]")
                 raise
 
 
 def check_tuple(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     # Specialized check for NamedTuples
     field_types = getattr(origin_type, "__annotations__", None)
@@ -324,7 +350,7 @@ def check_tuple(
 
         for name, field_type in field_types.items():
             try:
-                check_type_internal(getattr(value, name), field_type, memo)
+                check_type_internal(getattr(value, name), field_type, memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"attribute {name!r}")
                 raise
@@ -343,10 +369,10 @@ def check_tuple(
 
     if use_ellipsis:
         element_type = tuple_params[0]
-        samples = global_config.collection_check_strategy.iterate_samples(value)
+        samples = config.collection_check_strategy.iterate_samples(value)
         for i, element in enumerate(samples):
             try:
-                check_type_internal(element, element_type, memo)
+                check_type_internal(element, element_type, memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"item {i}")
                 raise
@@ -362,19 +388,23 @@ def check_tuple(
 
         for i, (element, element_type) in enumerate(zip(value, tuple_params)):
             try:
-                check_type_internal(element, element_type, memo)
+                check_type_internal(element, element_type, memo, config)
             except TypeCheckError as exc:
                 exc.append_path_element(f"item {i}")
                 raise
 
 
 def check_union(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     errors: dict[str, TypeCheckError] = {}
     for type_ in args:
         try:
-            check_type_internal(value, type_, memo)
+            check_type_internal(value, type_, memo, config)
             return
         except TypeCheckError as exc:
             errors[get_type_name(type_)] = exc
@@ -386,12 +416,16 @@ def check_union(
 
 
 def check_uniontype(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     errors: dict[str, TypeCheckError] = {}
     for type_ in args:
         try:
-            check_type_internal(value, type_, memo)
+            check_type_internal(value, type_, memo, config)
             return
         except TypeCheckError as exc:
             errors[get_type_name(type_)] = exc
@@ -403,7 +437,11 @@ def check_uniontype(
 
 
 def check_class(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isclass(value):
         raise TypeCheckError("is not a class")
@@ -416,9 +454,9 @@ def check_class(
     if expected_class is Any:
         return
     elif getattr(expected_class, "_is_protocol", False):
-        check_protocol(value, expected_class, (), memo)
+        check_protocol(value, expected_class, (), memo, config)
     elif isinstance(expected_class, TypeVar):
-        check_typevar(value, expected_class, (), memo, subclass_check=True)
+        check_typevar(value, expected_class, (), memo, config, subclass_check=True)
     elif get_origin(expected_class) is Union:
         errors: dict[str, TypeCheckError] = {}
         for arg in get_args(expected_class):
@@ -426,7 +464,7 @@ def check_class(
                 return
 
             try:
-                check_class(value, type, (arg,), memo)
+                check_class(value, type, (arg,), memo, config)
                 return
             except TypeCheckError as exc:
                 errors[get_type_name(arg)] = exc
@@ -442,7 +480,11 @@ def check_class(
 
 
 def check_newtype(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     supertype = origin_type.__supertype__
     if not isinstance(value, supertype):
@@ -450,7 +492,11 @@ def check_newtype(
 
 
 def check_instance(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isinstance(value, origin_type):
         raise TypeCheckError(f"is not an instance of {qualified_name(origin_type)}")
@@ -461,6 +507,7 @@ def check_typevar(
     origin_type: TypeVar,
     args: tuple[Any, ...],
     memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
     *,
     subclass_check: bool = False,
 ) -> None:
@@ -468,12 +515,12 @@ def check_typevar(
         annotation = (
             Type[origin_type.__bound__] if subclass_check else origin_type.__bound__
         )
-        check_type_internal(value, annotation, memo)
+        check_type_internal(value, annotation, memo, config)
     elif origin_type.__constraints__:
         for constraint in origin_type.__constraints__:
             annotation = Type[constraint] if subclass_check else constraint
             try:
-                check_type_internal(value, annotation, memo)
+                check_type_internal(value, annotation, memo, config)
             except TypeCheckError:
                 pass
             else:
@@ -488,7 +535,11 @@ def check_typevar(
 
 
 def check_literal(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     def get_literal_args(literal_args: tuple[Any, ...]) -> tuple[Any, ...]:
         retval: list[Any] = []
@@ -519,26 +570,42 @@ def check_literal(
 
 
 def check_literal_string(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
-    check_type_internal(value, str, memo)
+    check_type_internal(value, str, memo, config)
 
 
 def check_typeguard(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
-    check_type_internal(value, bool, memo)
+    check_type_internal(value, bool, memo, config)
 
 
 def check_none(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if value is not None:
         raise TypeCheckError("is not None")
 
 
 def check_number(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if origin_type is complex and not isinstance(value, (complex, float, int)):
         raise TypeCheckError("is neither complex, float or int")
@@ -547,7 +614,11 @@ def check_number(
 
 
 def check_io(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if origin_type is TextIO or (origin_type is IO and args == (str,)):
         if not isinstance(value, TextIOBase):
@@ -560,7 +631,11 @@ def check_io(
 
 
 def check_protocol(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     # TODO: implement proper compatibility checking and support non-runtime protocols
     if getattr(origin_type, "_is_runtime_protocol", False):
@@ -577,14 +652,22 @@ def check_protocol(
 
 
 def check_byteslike(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isinstance(value, (bytearray, bytes, memoryview)):
         raise TypeCheckError("is not bytes-like")
 
 
 def check_self(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if memo.self_type is None:
         raise TypeCheckError("cannot be checked against Self outside of a method call")
@@ -602,19 +685,32 @@ def check_self(
 
 
 def check_paramspec(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     pass  # No-op for now
 
 
 def check_instanceof(
-    value: Any, origin_type: Any, args: tuple[Any, ...], memo: TypeCheckMemo
+    value: Any,
+    origin_type: Any,
+    args: tuple[Any, ...],
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
 ) -> None:
     if not isinstance(value, origin_type):
         raise TypeCheckError(f"is not an instance of {qualified_name(origin_type)}")
 
 
-def check_type_internal(value: Any, annotation: Any, memo: TypeCheckMemo) -> None:
+def check_type_internal(
+    value: Any,
+    annotation: Any,
+    memo: TypeCheckMemo,
+    config: TypeCheckConfiguration,
+) -> None:
     """
     Check that the given object is compatible with the given type annotation.
 
@@ -631,9 +727,9 @@ def check_type_internal(value: Any, annotation: Any, memo: TypeCheckMemo) -> Non
         try:
             annotation = evaluate_forwardref(annotation, memo)
         except NameError:
-            if global_config.forward_ref_policy is ForwardRefPolicy.ERROR:
+            if config.forward_ref_policy is ForwardRefPolicy.ERROR:
                 raise
-            elif global_config.forward_ref_policy is ForwardRefPolicy.WARN:
+            elif config.forward_ref_policy is ForwardRefPolicy.WARN:
                 warnings.warn(
                     f"Cannot resolve forward reference {annotation.__forward_arg__!r}",
                     TypeHintWarning,
@@ -671,7 +767,7 @@ def check_type_internal(value: Any, annotation: Any, memo: TypeCheckMemo) -> Non
     for lookup_func in checker_lookup_functions:
         checker = lookup_func(origin_type, args, extras)
         if checker:
-            checker(value, origin_type, args, memo)
+            checker(value, origin_type, args, memo, config)
             return
 
     if not isinstance(value, origin_type):