+++ src/typeguard/_functions.py
@@ -8,7 +8,12 @@ from threading import Lock
 from typing import Any, Callable, Iterable, NoReturn, TypeVar, cast, overload
 
 from ._checkers import BINARY_MAGIC_METHODS, check_type_internal
-from ._config import global_config
+from ._config import (
+    CollectionCheckStrategy,
+    ForwardRefPolicy,
+    TypeCheckConfiguration,
+    global_config,
+)
 from ._exceptions import TypeCheckError, TypeCheckWarning
 from ._memo import TypeCheckMemo
 from ._utils import qualified_name
@@ -26,16 +31,41 @@ type_checks_suppress_lock = Lock()
 
 
 @overload
-def check_type(value: object, expected_type: type[T]) -> T:
+def check_type(
+    value: object,
+    expected_type: type[T],
+    *,
+    forward_ref_policy: ForwardRefPolicy = ...,
+    typecheck_fail_callback: TypeCheckFailCallback | None = ...,
+    collection_check_strategy: CollectionCheckStrategy = ...,
+) -> T:
     ...
 
 
 @overload
-def check_type(value: object, expected_type: Any) -> Any:
+def check_type(
+    value: object,
+    expected_type: Any,
+    *,
+    forward_ref_policy: ForwardRefPolicy = ...,
+    typecheck_fail_callback: TypeCheckFailCallback | None = ...,
+    collection_check_strategy: CollectionCheckStrategy = ...,
+) -> Any:
     ...
 
 
-def check_type(value: object, expected_type: Any) -> Any:
+def check_type(
+    value: object,
+    expected_type: Any,
+    *,
+    forward_ref_policy: ForwardRefPolicy = TypeCheckConfiguration().forward_ref_policy,
+    typecheck_fail_callback: (TypeCheckFailCallback | None) = (
+        TypeCheckConfiguration().typecheck_fail_callback
+    ),
+    collection_check_strategy: CollectionCheckStrategy = (
+        TypeCheckConfiguration().collection_check_strategy
+    ),
+) -> Any:
     """
     Ensure that ``value`` matches ``expected_type``.
 
@@ -49,23 +79,40 @@ def check_type(value: object, expected_type: Any) -> Any:
     * Forms a :class:`~.TypeCheckMemo` from the current stack frame
     * Calls the configured type check fail callback if the check fails
 
+    Note that this function is independent of the globally shared configuration in
+    :data:`typeguard.config`. This means that usage within libraries is safe from being
+    affected configuration changes made by other libraries or by the integrating
+    application. Instead, configuration options have the same default values as their
+    corresponding fields in :class:`TypeCheckConfiguration`.
+
     :param value: value to be checked against ``expected_type``
     :param expected_type: a class or generic type instance
+    :param forward_ref_policy: see :attr:`TypeCheckConfiguration.forward_ref_policy`
+    :param typecheck_fail_callback:
+        see :attr`TypeCheckConfiguration.typecheck_fail_callback`
+    :param collection_check_strategy:
+        see :attr:`TypeCheckConfiguration.collection_check_strategy`
     :return: ``value``, unmodified
     :raises TypeCheckError: if there is a type mismatch
 
     """
+    config = TypeCheckConfiguration(
+        forward_ref_policy=forward_ref_policy,
+        typecheck_fail_callback=typecheck_fail_callback,
+        collection_check_strategy=collection_check_strategy,
+    )
+
     if type_checks_suppressed or expected_type is Any:
         return
 
     frame = sys._getframe(1)
     memo = TypeCheckMemo(frame.f_globals, frame.f_locals)
     try:
-        check_type_internal(value, expected_type, memo)
+        check_type_internal(value, expected_type, memo, config)
     except TypeCheckError as exc:
         exc.append_path_element(qualified_name(value, add_class_prefix=True))
-        if global_config.typecheck_fail_callback:
-            global_config.typecheck_fail_callback(exc, memo)
+        if config.typecheck_fail_callback:
+            config.typecheck_fail_callback(exc, memo)
         else:
             raise
 
@@ -99,7 +146,7 @@ def check_argument_types(
                 raise exc
 
         try:
-            check_type_internal(value, annotation, memo=memo)
+            check_type_internal(value, annotation, memo, global_config)
         except TypeCheckError as exc:
             qualname = qualified_name(value, add_class_prefix=True)
             exc.append_path_element(f'argument "{argname}" ({qualname})')
@@ -125,7 +172,7 @@ def check_return_type(
             raise exc
 
     try:
-        check_type_internal(retval, annotation, memo)
+        check_type_internal(retval, annotation, memo, global_config)
     except TypeCheckError as exc:
         # Allow NotImplemented if this is a binary magic method (__eq__() et al)
         if retval is NotImplemented and annotation is bool:
@@ -160,7 +207,7 @@ def check_send_type(
             raise exc
 
     try:
-        check_type_internal(sendval, annotation, memo)
+        check_type_internal(sendval, annotation, memo, global_config)
     except TypeCheckError as exc:
         qualname = qualified_name(sendval, add_class_prefix=True)
         exc.append_path_element(f"the value sent to generator ({qualname})")
@@ -186,7 +233,7 @@ def check_yield_type(
             raise exc
 
     try:
-        check_type_internal(yieldval, annotation, memo)
+        check_type_internal(yieldval, annotation, memo, global_config)
     except TypeCheckError as exc:
         qualname = qualified_name(yieldval, add_class_prefix=True)
         exc.append_path_element(f"the yielded value ({qualname})")
@@ -215,7 +262,7 @@ def check_variable_assignment(
     ):
         iterated_values.append(obj)
         try:
-            check_type_internal(obj, expected_type, memo)
+            check_type_internal(obj, expected_type, memo, global_config)
         except TypeCheckError as exc:
             exc.append_path_element(argname)
             if global_config.typecheck_fail_callback: